separate Caller and loader
